web server & read on awake (no observe)

+++ leshan/leshan-client-demo/src/main/java/org/eclipse/leshan/client/demo/RandomGenericSensor.java
@@ -22,111 +22,111 @@ import org.slf4j.LoggerFactory;
 
 public class RandomGenericSensor extends BaseInstanceEnabler implements Destroyable {
 
-    private static final Logger LOG = LoggerFactory.getLogger(RandomGenericSensor.class);
+	private static final Logger LOG = LoggerFactory.getLogger(RandomGenericSensor.class);
 
-    private static final String UNIT_CONC = "ug/m3";
-    private static final int SENSOR_VALUE = 5700;
-    private static final int UNITS = 5701;
-    private static final int MIN_MEASURED_VALUE = 5601;
-    private static final int MAX_MEASURED_VALUE = 5602;
-    private static final int RESET_MIN_MAX_MEASURED_VALUES = 5605;
-    private static final int TIMESTAMP = 5518;
-    private static final List<Integer> supportedResources = Arrays.asList(SENSOR_VALUE, UNITS, MAX_MEASURED_VALUE,
-            MIN_MEASURED_VALUE, RESET_MIN_MAX_MEASURED_VALUES, TIMESTAMP);
-    private final ScheduledExecutorService scheduler;
-    private final Random rng = new Random();
-    private Date timestamp;
-    private double currentGen = 20d;
-    private double minMeasuredValue = currentGen;
-    private double maxMeasuredValue = currentGen;
+	private static final String UNIT_CONC = "ug/m3";
+	private static final int SENSOR_VALUE = 5700;
+	private static final int UNITS = 5701;
+	private static final int MIN_MEASURED_VALUE = 5601;
+	private static final int MAX_MEASURED_VALUE = 5602;
+	private static final int RESET_MIN_MAX_MEASURED_VALUES = 5605;
+	private static final int TIMESTAMP = 5518;
+	private static final List<Integer> supportedResources = Arrays.asList(SENSOR_VALUE, UNITS, MAX_MEASURED_VALUE,
+			MIN_MEASURED_VALUE, RESET_MIN_MAX_MEASURED_VALUES, TIMESTAMP);
+	private final ScheduledExecutorService scheduler;
+	private final Random rng = new Random();
+	private Date timestamp;
+	private float currentGen = 20f;
+	private float minMeasuredValue = currentGen;
+	private float maxMeasuredValue = currentGen;
 
-    public RandomGenericSensor() {
-        this.scheduler = Executors.newSingleThreadScheduledExecutor(new NamedThreadFactory("Generic Sensor"));
-        scheduler.scheduleAtFixedRate(new Runnable() {
+	public RandomGenericSensor() {
+		this.scheduler = Executors.newSingleThreadScheduledExecutor(new NamedThreadFactory("Generic Sensor"));
+		scheduler.scheduleAtFixedRate(new Runnable() {
 
-            @Override
-            public void run() {
-                adjustGeneric();
-            }
-        }, 2, 30, TimeUnit.SECONDS);
-    }
+			@Override
+			public void run() {
+				adjustGeneric();
+			}
+		}, 2, 30, TimeUnit.SECONDS);
+	}
 
-    @Override
-    public synchronized ReadResponse read(ServerIdentity identity, int resourceId) {
-        LOG.info("Read on Generic resource /{}/{}/{}", getModel().id, getId(), resourceId);
-        switch (resourceId) {
-        case MIN_MEASURED_VALUE:
-            return ReadResponse.success(resourceId, getTwoDigitValue(minMeasuredValue));
-        case MAX_MEASURED_VALUE:
-            return ReadResponse.success(resourceId, getTwoDigitValue(maxMeasuredValue));
-        case SENSOR_VALUE:
-            return ReadResponse.success(resourceId, getTwoDigitValue(currentGen));
-        case UNITS:
-            return ReadResponse.success(resourceId, UNIT_CONC);
-        case TIMESTAMP:
-            return ReadResponse.success(resourceId, getTimestamp());
-        default:
-            return super.read(identity, resourceId);
-        }
-    }
+	@Override
+	public synchronized ReadResponse read(ServerIdentity identity, int resourceId) {
+		LOG.info("Read on Generic resource /{}/{}/{}", getModel().id, getId(), resourceId);
+		switch (resourceId) {
+		case MIN_MEASURED_VALUE:
+			return ReadResponse.success(resourceId, getTwoDigitValue(minMeasuredValue));
+		case MAX_MEASURED_VALUE:
+			return ReadResponse.success(resourceId, getTwoDigitValue(maxMeasuredValue));
+		case SENSOR_VALUE:
+			return ReadResponse.success(resourceId, getTwoDigitValue(currentGen));
+		case UNITS:
+			return ReadResponse.success(resourceId, UNIT_CONC);
+		case TIMESTAMP:
+			return ReadResponse.success(resourceId, getTimestamp());
+		default:
+			return super.read(identity, resourceId);
+		}
+	}
 
-    @Override
-    public synchronized ExecuteResponse execute(ServerIdentity identity, int resourceId, String params) {
-        LOG.info("Execute on Generic resource /{}/{}/{}", getModel().id, getId(), resourceId);
-        switch (resourceId) {
-        case RESET_MIN_MAX_MEASURED_VALUES:
-            resetMinMaxMeasuredValues();
-            return ExecuteResponse.success();
-        default:
-            return super.execute(identity, resourceId, params);
-        }
-    }
+	@Override
+	public synchronized ExecuteResponse execute(ServerIdentity identity, int resourceId, String params) {
+		LOG.info("Execute on Generic resource /{}/{}/{}", getModel().id, getId(), resourceId);
+		switch (resourceId) {
+		case RESET_MIN_MAX_MEASURED_VALUES:
+			resetMinMaxMeasuredValues();
+			return ExecuteResponse.success();
+		default:
+			return super.execute(identity, resourceId, params);
+		}
+	}
 
-    private double getTwoDigitValue(double value) {
-        BigDecimal toBeTruncated = BigDecimal.valueOf(value);
-        return toBeTruncated.setScale(2, RoundingMode.HALF_UP).doubleValue();
-    }
+	private float getTwoDigitValue(float value) {
+		BigDecimal toBeTruncated = BigDecimal.valueOf(value);
+		return toBeTruncated.setScale(2, RoundingMode.HALF_UP).floatValue();
+	}
 
-    private void adjustGeneric() {
-        float delta = (rng.nextInt(20) - 10) / 10f;
-        currentGen += delta;
-        Integer changedResource = adjustMinMaxMeasuredValue(currentGen);
-        timestamp = new Date();
-        if (changedResource != null) {
-            fireResourcesChange(SENSOR_VALUE, changedResource);
-        } else {
-            fireResourcesChange(SENSOR_VALUE);
-        }
-    }
+	private void adjustGeneric() {
+		float delta = (rng.nextInt(20) - 10) / 10f;
+		currentGen += delta;
+		Integer changedResource = adjustMinMaxMeasuredValue(currentGen);
+		timestamp = new Date();
+		if (changedResource != null) {
+			fireResourcesChange(SENSOR_VALUE, changedResource);
+		} else {
+			fireResourcesChange(SENSOR_VALUE);
+		}
+	}
 
-    private synchronized Integer adjustMinMaxMeasuredValue(double newGeneric) {
-        if (newGeneric > maxMeasuredValue) {
-            maxMeasuredValue = newGeneric;
-            return MAX_MEASURED_VALUE;
-        } else if (newGeneric < minMeasuredValue) {
-            minMeasuredValue = newGeneric;
-            return MIN_MEASURED_VALUE;
-        } else {
-            return null;
-        }
-    }
+	private synchronized Integer adjustMinMaxMeasuredValue(float newGeneric) {
+		if (newGeneric > maxMeasuredValue) {
+			maxMeasuredValue = newGeneric;
+			return MAX_MEASURED_VALUE;
+		} else if (newGeneric < minMeasuredValue) {
+			minMeasuredValue = newGeneric;
+			return MIN_MEASURED_VALUE;
+		} else {
+			return null;
+		}
+	}
 
-    private void resetMinMaxMeasuredValues() {
-        minMeasuredValue = currentGen;
-        maxMeasuredValue = currentGen;
-    }
+	private void resetMinMaxMeasuredValues() {
+		minMeasuredValue = currentGen;
+		maxMeasuredValue = currentGen;
+	}
 
-    private Date getTimestamp() {
-        return timestamp;
-    }
+	private Date getTimestamp() {
+		return timestamp;
+	}
 
-    @Override
-    public List<Integer> getAvailableResourceIds(ObjectModel model) {
-        return supportedResources;
-    }
+	@Override
+	public List<Integer> getAvailableResourceIds(ObjectModel model) {
+		return supportedResources;
+	}
 
-    @Override
-    public void destroy() {
-        scheduler.shutdown();
-    }
+	@Override
+	public void destroy() {
+		scheduler.shutdown();
+	}
 }
